% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_normalizer.R
\name{address_normalizer}
\alias{address_normalizer}
\title{Címek tisztítása}
\usage{
address_normalizer(x)
}
\arguments{
\item{x}{tisztítandó címeket tartalmazó vektor}
}
\value{
sztring vektor, \code{x} minden elemére egy javított verzió
}
\description{
Egy irányítószámból, településből, köztérből és házszámból álló
  címek egységes alakra hozása.
}

