% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanCity.R
\name{cleanCity}
\alias{cleanCity}
\title{Irányítószám tisztítása}
\usage{
cleanCity(x)
}
\arguments{
\item{x}{karakter vektor}
}
\value{
\code{x}-nel azonos hosszú karakter vektor
}
\description{
Megtisztítja a kapott városneveket potenciális problémáktól.
}
\examples{
cleanZip(" 3120 Dorgalfalva, Kossuth utca 10.")

}
\author{
Hajdú László
}

