% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDates.R
\name{cleanDates}
\alias{cleanDates}
\title{Dátumtisztíítás}
\usage{
cleanDates(chr_vect, correct_yy_flag = TRUE)
}
\arguments{
\item{chr_vect}{karakter vektor}

\item{correct_yy_flag}{logikau érték, készámjegyű évszámok kezelésére.
Alapvetően \code{TRUE}}
}
\value{
\code{x}-szel megegyező hosszú Dátum vektor
}
\description{
Dátumok formátumának ellenőrzése. Erre lehet, hogy a \code{libridate}
  csomag eleve jobb.
}
\details{
A függvény megkeresi pár előre ismert formátumú dárumokat, de gőzöm
  sincs mit csinál velük.
A \code{correct_yy_flag} a két évszámból álló éveket kiegészíti:
  ha nagyobb, mint az aktuáis év utolsó két szánjegye, feltételezi,
  hogy a XX. századra vonatkozik.
}
\examples{
cleanDates("56-10-23")

}
\author{
Tajti András
}

