% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanEmails.R
\encoding{UTF-8}
\name{cleanEmails}
\alias{cleanEmails}
\title{E-mail címek tisztítása}
\usage{
cleanEmails(chr_vect)
}
\arguments{
\item{chr_vect}{karakter vektor}
}
\value{
\code{x}-szel megegyező hosszú karakter vektor
}
\description{
A nem megfelelő formátumú e-ail-címeket (amikben nem csak egy
  \code{@} és \code{.} van), \code{NA}-val helyettesíti.
}
\details{
A függvény jelenleg egy \code{grepl}-hívásból áll. Találtunk duplikált
  is, ennek detektálása a két \code{@} és \code{.}, vagy hogy minden
  előfordulásának száma páros. Fel lehet később turbózni, hogy
  létező internetkapcsolat esetén ellenőrizze hogy létező top level
  domain-nel rendelkezik-e a cím.
}
\examples{
cleanEmails("vmi@email.cim")

}
\author{
Tomasovszky Álmos
}
