% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFirms.R
\name{cleanFirms}
\alias{cleanFirms}
\title{Cégnevek tisztítása}
\usage{
cleanFirms(x)
}
\arguments{
\item{x}{egyszerű tisztításon átesett karakter vektor}
}
\description{
Nagybetűs, ékezetmentes cégnevekben keres gyakran előforduló elemeket
  a függvény, amiket egységes alakra hoz
}
\section{TODO}{

Javítani, finomítani a keresendő mintákat
}
\examples{
cleanFirms(toupper(unAccent(c("Andego Tanácsadó Korlátolt felelősségű társaság",
                              "Magyar Telekom Nyrt"))))

}
\author{
András Tajti <atajti@andego.hu>
}

