% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNames.R
\name{cleanNames}
\alias{cleanNames}
\title{Névtisztítás}
\usage{
cleanNames(x)
}
\arguments{
\item{x}{karakter vektor, megtisztítandó személynevekkel}
}
\value{
karakter vektor, a tisztított nevekkel
}
\description{
Cég- és személynevek egységes alakra hozása
}
\details{
A függvény valójában csak egy wrapper, a \code{detectFirms} eredménye
  alapján eldönti, melyik névre alkalmazza a \code{cleanPersonNames}
  és melyikre a \code{cleanFirms} függvényt.
}
\examples{
cleanNames("Sándor József Benedek")
cleanNames("S. József Benedek")
cleanNames("Sándor J. Benedek")
cleanNames(c("Andego Tanácsadó Korlátolt felelősségű társaság",
             "Magyar Telekom Nyrt"))

}
\author{
Tajti András <atajti@andego.hu>
}

