% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPersonNames.R
\name{cleanPersonNames}
\alias{cleanPersonNames}
\title{Névtisztítás}
\usage{
cleanPersonNames(x)
}
\arguments{
\item{x}{karakter vektor, megtisztítandó személynevekkel}
}
\value{
karakter vektor, a tisztított nevekkel
}
\description{
Személynevek egységes alapkra hozása
}
\details{
Alapvetően vezetéknévre és keresztnévre bontjuk a neveket.
Ha van középs név, azt az első betűre redukáljuk (Kiss Sándor József
  -> Kiss S József'), kivétel, ha az első név egy vagy két betűből
  áll - ekkor a középső nevet teljesen meghagyjuk (N Nagy Nándor)
}
\section{TODO}{

Egyrészt karaktereket és nem betűket különböztetünk meg, Szabó Sz
  Balázs -> Szabó S Balázs, másrészt római számokat is érdemes
  lesz figyelni később: II. Rákóczi Ferenc -> I Rákóczi Ferenc
}
\examples{
cleanPersonNames("Sándor József Benedek")
cleanPersonNames("S. József Benedek")
cleanPersonNames("Sándor J. Benedek")

}
\author{
Tajti András <atajti@andego.hu>
}

