% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPhones.R
\name{cleanPhones}
\alias{cleanPhones}
\title{Telefonszámok tisztítása}
\usage{
cleanPhones(chr_vect)
}
\arguments{
\item{chr_vect}{karakter vektor}
}
\value{
\code{chr_vect}-tel megegyező hosszú karakter vektor
}
\description{
Magyar telefonszámokra beállított telefonszámtisztítás
}
\details{
Első lépésként csak a számokat tartja meg a szövegből. Amennyiben
  kevesebb mint hét számjegyet talál, illetve ha üres sztringet talál,
  figyelmeztet, és az értéket \code{NA}-ra állítja.
}
\section{TODO}{

Internacionalizálás: egy \code{lang} argumentum, ami ha meg van adva,
  megkeresné a \code{paste0(cleanPhones_*lang*)} függvényt, és annak
  a vektort.
}

\examples{
cleanPhones("+36 (30) 1010101")

}
\author{
Tomasovszky Álmos
}
