% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanStreet.R
\name{cleanStreet}
\alias{cleanStreet}
\title{Utcanevek tisztítása}
\usage{
cleanStreet(x)
}
\arguments{
\item{x}{karakter vektor}
}
\value{
\code{x}-nel azonos hosszú karakter vektor
}
\description{
Megtisztítja a kapott irányítószámokat potenciális problémáktól.
}
\section{TODO}{

Jelenleg a gyakori hibás utcanevek a kódba vannak égetve,
  ezt később érdemes volna modulárisan cserélhetővé tenni.

Jelenleg az \code{unAccent(x, alternative=TRUE)} van használatban,
  mivel az \code{iconv} a rendszer kódolását, és nem a sztring
  kódolását veszi alapul
}
\examples{
cleanStreet(" 3120 Dorgalfalva, Kossuth utca 10.")

}
\author{
Hajdú László
}

