% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constReplacer.R
\encoding{UTF-8}
\name{constReplacer}
\alias{constReplacer}
\title{Konstansokban tárolt elemek cseréje}
\usage{
constReplacer(regkif, csere, min)
}
\arguments{
\item{regkif}{reguláris kifejezés, u.a. mint a \code{pattern}}

\item{csere}{cseresztring, u.a. mint a \code{replacement}}

\item{min}{a minta amiben keres, u.a. mint az \code{x}'}
}
\value{
\code{min}-nel azonos hosszú karakter vektor
}
\description{
Csak eszköz, valójában egy \code{gsub}-hívás.
}
\author{
Hajdú László
}
