% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectFirms.R
\name{detectFirms}
\alias{detectFirms}
\title{Cégek keresése}
\usage{
detectFirms(chr_vect)
}
\arguments{
\item{chr_vect}{egyszerű tisztításon átesett karakter vektor}
}
\value{
logikai vektor, TRUE ha cég, FALSE ha személy.
}
\description{
Nagybetűs, ékezetmentes nevekben keres gyakran előforduló elemeket
  a függvény, amik cégnévre utalnak
}
\section{TODO}{

Javítani, finomítani a keresendő mintákat
}

\examples{
cleanFirms(toupper(unAccent(c("Andego Tanácsadó Korlátolt felelősségű társaság",
                              "Magyar Telekom Nyrt",
                              "Tajti András"))))
}
\author{
András Tajti <atajti@andego.hu>
}
