% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSpaces.R
\encoding{UTF-8}
\name{removeSpaces}
\alias{removeSpaces}
\title{Felesleges szóközök eltávolítása}
\usage{
removeSpaces(x)
}
\arguments{
\item{x}{karakter vektor}
}
\value{
\code{x}-szel megegyező hosszú karakter vektor
}
\description{
Ha több mint egy sortörés, tab, kocsivissza vagy szóköz jön egymás
után, egyetlen szóközre cseréli, illetve a szöveg eleji és végi
szóközöket is eltüntetni.
}
\details{
\code{trimws(gsub("[ \n\t\r]+", "", x))}
}
\section{TODO}{

A cél az lenne, hogy minden whitespace-t csak egy space-re
cseréljünk?
}

\examples{
removeSpaces("Az ellipszis argumentum: ...")

}
\author{
Hajdu László
}
