% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSpecials.R
\name{removeSpecials}
\alias{removeSpecials}
\title{Speciális karakterek eltűntetése}
\usage{
removeSpecials(x)
}
\arguments{
\item{x}{karakter vektor}
}
\value{
\code{x}-szel megegyező hosszú karakter vektor
}
\description{
Csak betűk, számok és üres karaktereket tart meg, a többit kivágja.
}
\details{
Az :alnum: és a :blank: osztályú karakterek maradnak csak meg.
}
\examples{
removeSpecials("Uj\\nsor")

}
\author{
Hajdu László
}

