% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toAlphaAndDigits.R
\name{toAlphaAndDigits}
\alias{toAlphaAndDigits}
\title{Számok és betűk megtartása}
\usage{
toAlphaAndDigits(x)
}
\arguments{
\item{x}{karakter vektor}
}
\value{
\code{x}-szel megegyező hosszú karakter vektor
}
\description{
Minden karaktert töröl, ami nem szám, vagy nem betű (a betű
  \code{:alpha:} értelmében, ld. \code{?regex})
}
\details{
\code{gsub("[0-9]","",x)}
}
\section{TODO}{

A nem angolszász ABC betűit is békén hagyhatná...
}
\examples{
toAlphaAndDigits("1+4=öt")

}
\author{
Hajdu László
}

