% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toZipCode.R
\name{toZipCode}
\alias{toZipCode}
\title{Irányítószám kiszűrése}
\usage{
toZipCode(x)
}
\arguments{
\item{x}{karakter vektor}
}
\value{
\code{x}-szel megegyező hosszú karakter vektor
}
\description{
Az első 4 egymás utáni számot tartja meg.
}
\details{
\code{str_extract(x,"\\b[0-9]{4}\\b")}
}
\examples{
cimek <- c("8000 Siófok, Ballagó utca 14/a",
           "8000 Siofk, Ballagó utca 14/a 3. em 31.")
toZipCode(cimek)

}
\author{
Hajdu László
}

