% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unAccent.R
\encoding{UTF-8}
\name{unAccent}
\alias{unAccent}
\title{Ékezetek eltávolítása}
\usage{
unAccent(x, alternative = FALSE)
}
\arguments{
\item{x}{karakter vektor}

\item{alternative}{logikai konstans, alapvetően hamis. Ha igaz,
akkor egy beépített cserekészlettel dolgozik kódolás helyett.}
}
\value{
\code{x}-szel egyező hosszú karakter vektor, ékezetek nélkül.
}
\description{
Az ékezetes karaktereket (pl á,é,í) ékezet nélkülire cseréli (a,e,i).
}
\details{
Az ékezet eltávolítását az \code{iconv(x, from="", to="ASCII//TRANSLIT")}
  függvény csinálja.
A \code{factor} változókat feldolgozás előtt karakterré alakítja
}
\examples{
unAccent("Árvíztűrő tükörfúrógép")
unAccent("Árvíztűrő tükörfúrógép", TRUE)

}
\author{
Hajdú László, Tajti András
}
