% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFirms.R
\name{cleanFirms}
\alias{cleanFirms}
\title{Cégnevek tisztítása}
\usage{
cleanFirms(x)
}
\arguments{
\item{x}{karakter vektor}
}
\description{
A cégneveket egységesíti, pl.: NyRT -> RT
}
\section{TODO}{

Javítani, finomítani a keresendő mintákat
}

\examples{
cleanFirms(c("Andego Tanácsadó Korlátolt felelősségű társaság",
             "Magyar Telekom Nyrt"))

}
\author{
András Tajti <atajti@andego.hu>
}
