% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPersonNames.R
\name{cleanPersonNames}
\alias{cleanPersonNames}
\title{Névtisztítás}
\usage{
cleanPersonNames(x)
}
\arguments{
\item{x}{karakter vektor, megtisztítandó személynevekkel}
}
\value{
karakter vektor, a tisztított nevekkel
}
\description{
Személynevek egységes alapkra hozása
}
\details{
A nevej első két tagja marad meg, kivétel, ha az első tag tisztított
  formában max 2 karakter hosszú (ld. példa), akkor az első három
  tag marad meg.
}
\section{TODO}{

Egyrészt karaktereket és nem betűket különböztetünk meg, Szabó Sz
  Balázs -> Szabó S Balázs, másrészt római számokat is érdemes
  lesz figyelni később: II. Rákóczi Ferenc -> I Rákóczi Ferenc
}

\examples{
cleanPersonNames("Sándor József Benedek")
cleanPersonNames("S. József Benedek")
cleanPersonNames("Sándor J. Benedek")

}
\author{
Tajti András <atajti@andego.hu>
}
