% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanStreet.R
\name{cleanStreet}
\alias{cleanStreet}
\title{Utcanevek tisztítása}
\usage{
cleanStreet(x, alternative = FALSE)
}
\arguments{
\item{x}{karakter vektor}

\item{alternative}{unAccent használatát szabályzó logikai érték}
}
\value{
\code{x}-szel azonos hosszú karakter vektor
}
\description{
Megtisztítja a kapott irányítószámokat potenciális problémáktól.
}
\details{
Az \code{unAccent} alapvetően az \code{iconv} függvényt használja.
  Ez az operácoós rendszer kódolását veszi alapul, így ha
  \code{Encoding(x)} nem ugyan az, mint ami az operációs rendszerből
  következik, hibás eredményt adhat; erre az esetre használható
  az \code{alternative} flag, lásd \code{\link{unAccent}}
}
\section{TODO}{

jelenleg csak szabályos, teljes címekre működik, csak címekre nem;
pl.: \code{cleanStreet("1000 Budapest, Clark Ádám tér 3")} működik,
  de \code{cleanStreet(Clark Ádám tér 3)} nem.

Jelenleg a gyakori hibás utcanevek a kódba vannak égetve,
  ezt később érdemes volna modulárisan cserélhetővé tenni.
}

\examples{
cleanStreet(" 3120 Dorgalfalva, Kossuth utca 10.")

}
\author{
Hajdú László
}
