% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanZip.R
\name{cleanZip}
\alias{cleanZip}
\title{Irányítószám tisztítása}
\usage{
cleanZip(x)
}
\arguments{
\item{x}{karakter vektor}
}
\value{
\code{x}-nel azonos hosszú karakter vektor
}
\description{
Megtisztítja a kapott irányítószámokat potenciális problémáktól.
}
\examples{
cleanZip(" 3120 Dorgalfalva, Kossuth utca 10.")

}
\author{
Hajdú László
}

